+++ froi/utils.py
@@ -2,7 +2,6 @@
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 
 import os
-import froi
 
 
 def get_file_names(path, postfix):
@@ -25,11 +24,11 @@ def get_icon_dir():
     """Get directory which contains icons."""
     pjoin = os.path.join
     apath = os.path.abspath
-    froi_dir = os.path.dirname(froi.__file__)
+    froi_dir = os.path.dirname(__file__)
     base_dir = apath(pjoin(froi_dir, os.pardir))
-    base_dir = apath(pjoin(base_dir, os.pardir))
+    #base_dir = apath(pjoin(base_dir, os.pardir))
     if not os.path.exists(pjoin(base_dir, 'data')):
-        icon_dir = pjoin(froi_dir,'gui', 'icon')
+        icon_dir = pjoin(froi_dir, 'icon')
     else:
         icon_dir = pjoin(base_dir, 'icon')
     return icon_dir
@@ -38,9 +37,9 @@ def get_data_dir():
     """Get data directory path."""
     pjoin = os.path.join
     apath = os.path.abspath
-    froi_dir = os.path.dirname(froi.__file__)
+    froi_dir = os.path.dirname(__file__)
     base_dir = apath(pjoin(froi_dir, os.pardir))
-    base_dir = apath(pjoin(base_dir, os.pardir))
+    #base_dir = apath(pjoin(base_dir, os.pardir))
     if not os.path.exists(pjoin(base_dir, 'data')):
         data_dir = pjoin(froi_dir,'data')
     else:
