+++ froi/widgets/binarizationdialog.py
@@ -5,7 +5,7 @@ __author__ = 'zhouguangfu'
 from PyQt4.QtCore import *
 from PyQt4.QtGui import *
 
-from froi.algorithm import imtool
+from ..algorithm import imtool
 
 
 class BinarizationDialog(QDialog):
