+++ setup.py
@@ -68,19 +68,20 @@ from version import __version__
 setup(name='freeroi',
       version=__version__,
       description='Toolbox for ROI defining',
-      author='Lijie Huang, Zetian Yang, Guangfu Zhou and Zonglei Zhen, from Neuroinformatic Team@LiuLab',
+      author='Lijie Huang, Zetian Yang, Guangfu Zhou, Zhaoguo Liu, and Zonglei Zhen, from Neuroinformatic Team@LiuLab',
       author_email=['huanglijie.seal@gmail.com', 'zetian.yang@gmail.com'],
       packages=['froi',
                 'froi.algorithm',
-                'froi.gui',
-                'froi.gui.base',
-                'froi.gui.component',
+                'froi.core',
+                'froi.widgets',
+                'froi.interface',
                 'froi.io'],
       scripts=['bin/freeroi', 'bin/freeroi-sess'],
       package_data={'froi': ['data/label/face-object/*',
-                             'data/labelconfig/face.lbl',
-                             'data/standard/MNI152_T1_2mm_brain.nii.gz',
-                             'gui/icon/*']
+                             'data/labelconfig/*',
+                             'data/standard/*',
+                             'data/atlas/*',
+                             'icon/*']
                     },
       ext_modules=[qimageview],
       cmdclass={'build_ext': build_ext}
